+++ src/state/sidebar-state.ts
@@ -1,10 +1,11 @@
+import type { WeeklyData } from '../task-timer/classes/report-info'
 import TaskTimer from '../task-timer/task-timer'
 import { formatDate } from '../task-timer/utils/date-utils'
 import * as vscode from 'vscode'
 
 export default class SidebarState {
   calendarDate: Date | null = null
-  pageContents: string | null = null
+  pageData: WeeklyData | null = null
 
   async setCalendarDate (date: Date): Promise<void> {
     this.calendarDate = date
@@ -16,8 +17,8 @@ export default class SidebarState {
   private readonly _onDateChange = new vscode.EventEmitter<Date>()
   readonly onDateChange: vscode.Event<Date> = this._onDateChange.event
 
-  private readonly _onDailyContentsChange = new vscode.EventEmitter<string>()
-  readonly onDailyContentsChange: vscode.Event<string> = this._onDailyContentsChange.event
+  private readonly _onWeeklyDataChange = new vscode.EventEmitter<WeeklyData>()
+  readonly onWeeklyDataChange: vscode.Event<WeeklyData> = this._onWeeklyDataChange.event
 
   private readonly _openValueChange = new vscode.EventEmitter<boolean>()
   readonly onOpenValueChange: vscode.Event<boolean> = this._openValueChange.event
@@ -44,8 +45,8 @@ export default class SidebarState {
     }
 
     const timer = new TaskTimer(this.rootFilePath)
-    const contents = await timer.getDateContents(formatDate(this.calendarDate))
-    this.pageContents = contents
-    this._onDailyContentsChange.fire(contents ?? '')
+    const weeklyData = await timer.getWeeklyData(formatDate(this.calendarDate))
+    this.pageData = weeklyData
+    this._onWeeklyDataChange.fire(weeklyData)
   }
 }